requirements.txt
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
streamlit-ketcher
stmol>=0.0.9
py3Dmol>=2.0.0
mols2grid>=0.2.1

mordred
chembl-webresource-client>=0.10.9
pubchempy>=1.0.4
openmm>=8.0.0
mdanalysis>=2.4.0
scikit-learn>=1.3.0
torch>=2.0.0
plotly>=5.15.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0
biopython>=1.81
nglview>=3.0.3
ipywidgets
rdkit-pypi!=3.4.0